requirements.txt
asgiref==3.4.1
Django==3.2.5
django-3-jet==1.0.8
django-utils-six==2.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.1